server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cway-mcp-server"
version = "0.1.0"
description = "MCP server for Cway GraphQL API integration"
authors = [{name = "Fredrik Hultin"}]
requires-python = ">=3.9"
dependencies = [
    "mcp>=1.0.0",
    "gql[all]>=3.5.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-socketio>=5.14.0",
]

[project.scripts]
cway-mcp-server = "src.presentation.cway_mcp_server:main"
cway-server-with-dashboard = "start_server_with_dashboard:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "mypy>=1.6.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "httpx>=0.25.0",
    "pytest-mock>=3.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "gql.*",
    "mcp.*",
]
ignore_missing_imports = true
